refactor: move incoming option type declaration to options.ts file
